requirements.txt
matplotlib
mapclassify
rtree
gunicorn==20.1.0
folium==0.12.1.post1
geopandas==0.10.2
python-dotenv==0.21.0
schedule==1.1.0
Flask==2.2.2